Refine "view change log" options
Reworked the "view CHANGELOG" option on the debug menu. Now check if APK Manager was installed via git, and if so, parse and show the 15 most recent commits from the git log by default instead of opening the wiki page.
Added options to open the wiki page or the github commit history page.